# cmd/api/static/openapi.yaml
openapi: 3.0.3
info:
  title: DriveBai API
  description: |
    DriveBai - Car Marketplace and Rental App API

    ## Authentication
    Most endpoints require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```

    ## Error Handling
    All errors return a JSON object with an `error` field containing `code` and `message`.

    Error codes:
    - `EMAIL_TAKEN` - Email already registered
    - `INVALID_CREDENTIALS` - Wrong email or password
    - `RATE_LIMITED` - Too many requests
    - `UNAUTHORIZED` - Authentication required
    - `TOKEN_EXPIRED` - Access token expired
    - `TOKEN_INVALID` - Invalid token
    - `INVALID_ROLE` - Invalid role specified
    - `INCOMPLETE_ONBOARDING` - Onboarding requirements not met
  version: 0.2.0
  contact:
    name: DriveBai Support
    email: support@drivebai.com
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.drivebai.com/api/v1
    description: Production server
...
paths:
...
  /me:
    get:
      tags: [User]
      summary: Get current user profile
      description: Returns the authenticated user's profile information.
      operationId: getCurrentUser
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /profile:
    patch:
      tags: [User]
      summary: Update user profile
      description: |
        Updates the authenticated user's profile information.
        All fields are optional - only provided fields will be updated.
        Role can only be set to 'driver' or 'car_owner'.
      operationId: updateProfile
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdateProfileResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /profile/photo:
    post:
      tags: [User]
      summary: Upload profile photo
      description: Uploads a profile photo for the authenticated user.
      operationId: uploadProfilePhoto
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Image file (JPEG, PNG)
      responses:
        '200':
          description: Photo uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  profile_photo_url:
                    type: string
        '400':
          description: Invalid file type or size
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /documents:
    get:
      tags: [Documents]
      summary: Get user's documents
      description: Returns all documents uploaded by the authenticated user.
      operationId: getDocuments
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Document'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /documents/{type}:
    post:
      tags: [Documents]
      summary: Upload a document
      description: |
        Uploads a document of the specified type.
        If a document of the same type already exists, it will be replaced.
      operationId: uploadDocument
      security:
        - BearerAuth: []
      parameters:
        - name: type
          in: path
          required: true
          schema:
            type: string
            enum: [drivers_license, registration]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: Document file (JPEG, PNG, PDF)
      responses:
        '201':
          description: Document uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
        '400':
          description: Invalid file type or document type
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /documents/{id}:
    delete:
      tags: [Documents]
      summary: Delete a document
      description: Deletes the specified document.
      operationId: deleteDocument
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Document deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          description: Document not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /onboarding/complete:
    post:
      tags: [Onboarding]
      summary: Complete onboarding
      description: |
        Marks the user's onboarding as complete.
        For drivers, required documents must be uploaded first.
      operationId: completeOnboarding
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Onboarding completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Onboarding requirements not met
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
          enum: [driver, car_owner, admin]
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
          nullable: true
        is_email_verified:
          type: boolean
        onboarding_status:
          type: string
          enum: [created, role_selected, photo_uploaded, documents_uploaded, complete]
        profile_photo_url:
          type: string
          nullable: true
    UpdateProfileRequest:
      type: object
      properties:
        role:
          type: string
          enum: [driver, car_owner]
          description: User role (cannot be set to admin via API)
        first_name:
          type: string
        last_name:
          type: string
        phone:
          type: string
          nullable: true
    UpdateProfileResponse:
      type: object
      properties:
        message:
          type: string
        data:
          $ref: '#/components/schemas/UserProfile'
    Document:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [drivers_license, registration]
        file_name:
          type: string
        file_size:
          type: integer
        status:
          type: string
          enum: [uploaded, verified, rejected]
        created_at:
          type: string
          format: date-time
    MessageResponse:
      type: object
      properties:
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: INVALID_INPUT
            message:
              type: string
              example: Email is required